+++ src/kabr_tools/utils/detector.py
