+++ src/kabr_tools/utils/object.py
